Extract shared primitives to common.py to decouple V2 from V1
Moves `CoReasonBaseModel`, `StrictUri`, and `ToolRiskLevel` to a new `src/coreason_manifest/common.py` module.
Refactors V1 definitions (`base.py`, `agent.py`) and V2 specs (`v2/spec/definitions.py`) to import from this new location.
Updates all internal references to use the canonical location in `common.py`.
This eliminates the hard dependency of V2 on V1 legacy code, facilitating future deprecation of V1.
Backward compatibility is maintained as `base.py` and `agent.py` re-export the moved symbols (implicitly or explicitly).
