+++ src/coreason_manifest/definitions/audit.py
@@ -18,7 +18,7 @@ from uuid import UUID
 
 from pydantic import ConfigDict, Field, model_validator
 
-from coreason_manifest.definitions.base import CoReasonBaseModel
+from coreason_manifest.common import CoReasonBaseModel
 
 from .message import ChatMessage
