+++ src/coreason_manifest/definitions/contracts.py
@@ -15,7 +15,7 @@ from uuid import UUID
 from pydantic import ConfigDict, Field
 
 from coreason_manifest.definitions.agent import VersionStr
-from coreason_manifest.definitions.base import CoReasonBaseModel
+from coreason_manifest.common import CoReasonBaseModel
 
 
 class ContractMetadata(CoReasonBaseModel):
