+++ src/coreason_manifest/definitions/deployment.py
@@ -13,7 +13,7 @@ from typing import Dict
 
 from pydantic import ConfigDict, Field
 
-from coreason_manifest.definitions.base import CoReasonBaseModel
+from coreason_manifest.common import CoReasonBaseModel
 
 
 class Protocol(str, Enum):
