+++ src/coreason_manifest/definitions/message.py
@@ -16,7 +16,7 @@ from typing import Annotated, Any, Dict, List, Literal, Optional, Union, cast
 
 from pydantic import ConfigDict, Field, model_validator
 
-from coreason_manifest.definitions.base import CoReasonBaseModel
+from coreason_manifest.common import CoReasonBaseModel
 
 # --- Enums ---
