+++ src/coreason_manifest/definitions/session.py
@@ -14,7 +14,7 @@ from uuid import UUID, uuid4
 
 from pydantic import ConfigDict, Field
 
-from coreason_manifest.definitions.base import CoReasonBaseModel
+from coreason_manifest.common import CoReasonBaseModel
 from coreason_manifest.definitions.events import GraphEvent
 from coreason_manifest.definitions.identity import Identity
 from coreason_manifest.definitions.message import MultiModalInput
