+++ src/coreason_manifest/recipes.py
@@ -23,7 +23,7 @@ from typing import Any, Dict, Optional
 from pydantic import ConfigDict, Field
 
 from .definitions.agent import VersionStr
-from .definitions.base import CoReasonBaseModel
+from coreason_manifest.common import CoReasonBaseModel
 from .definitions.topology import GraphTopology, StateDefinition
